Build clean under -Wconversion, -Wsign-conversion, -Wshadow and -Wold-style-cast
grad_enable_warnings now adds all four for GCC and Clang (Clang gets
-Wshadow-all, which, like GCC's -Wshadow, also flags constructor
parameters and lambda parameters that shadow). Enabling them on main
reported 552 distinct warnings; the tree now builds with none, library,
CLI and tests alike, under -Werror.

Most came from size_t shapes meeting int interfaces, so the fix is at
the interfaces rather than at the call sites, under one convention now
written down in tensor.h: extents, counts, offsets and indices are
size_t; hyperparameters (d_model, num_heads, vocab_size, seq_length,
batch_size, step counts) stay int, since they come from flags, presets
and the int32 checkpoint header, and each module validates them and
converts once where they enter; token ids are int.

- blas_wrapper.h: blas_sgemm_ex, the vec_* helpers and metal::sgemm take
  size_t. The BLAS integer type is read off cblas_sdot's signature
  (long under Accelerate's ILP64, int for OpenBLAS), and GEMM arguments
  narrow to it once per call through the new grad::narrow, which throws
  rather than wraps. The per-row vec_* helpers only assert, keeping the
  kernels' inner loops free of new checks; Tensor's kMaxTensorElements
  (static_assert below INT_MAX) bounds every length they see.
- Attention's per-head cblas_sgemm calls go through a new
  blas_sgemm_strided, and the KV-cache session's sdot/saxpy through
  vec_dot/vec_axpy, so no kernel calls CBLAS directly.
- Attention, LayerNorm, embeddings, the KV-cache session, the data
  loader and datasets take their extents from tensor shapes as size_t;
  the int hyperparameter members gain a trailing underscore, matching
  rope_ and gated_, so constructor parameters no longer shadow them.
- Implicit int-to-float and double-to-float conversions are spelled
  out with the same operands, so every value is unchanged. The one
  place that needed care is AdamW's bias correction: pow(float, int)
  returns double, and the subtraction must stay in double before the
  single narrowing, exactly as before.
- Lambda parameters and locals that shadowed an enclosing local are
  renamed. The one C-style cast (task_info_t) becomes a
  reinterpret_cast, and the MAP_FAILED comparison is fenced with a
  pragma, since some GCCs report -Wold-style-cast through that macro.

GCC could not be run here. An AST pass over every translation unit
covered the conversions Clang accepts through range analysis but GCC
flags (it turned up 0x2800 + v in the dashboard, now unsigned), and
the Linux-only branches were reviewed by hand (the /proc statm reader
and termios flags).

Determinism probe, train-fast metrics, greedy decoding and val loss
1.6929 are bit-identical.

+++ include/grad/data/dataloader.h
@@ -10,7 +10,7 @@ struct Batch {
     Tensor input;
     Tensor target;
     
-    Batch(int batch_size, int seq_length)
+    Batch(size_t batch_size, size_t seq_length)
         : input(batch_size, seq_length, 1),
           target(batch_size, seq_length, 1) {}
 };
@@ -24,7 +24,7 @@ struct Batch {
 class DataLoader {
 private:
     std::shared_ptr<Dataset> dataset_;
-    int batch_size_;
+    size_t batch_size_;
     bool shuffle_;
     size_t current_index_;
     std::mt19937 rng_;
@@ -45,7 +45,8 @@ public:
     }
 
     const std::shared_ptr<Dataset>& dataset() const { return dataset_; }
-    int batch_size() const { return batch_size_; }
+    // batch_size was validated >= 1 as an int, so it converts back.
+    int batch_size() const { return static_cast<int>(batch_size_); }
 };
 
 }  // namespace grad
